+++ solomon-file/src/main/java/com/steven/solomon/service/COSService.java
@@ -73,6 +73,10 @@ public class COSService implements FileServiceInterface{
 
   @Override
   public void deleteFile(String fileName, String bucketName) throws Exception {
+    boolean flag = bucketExists(bucketName);
+    if (!flag) {
+      return;
+    }
     String filePath = getFilePath(fileName,properties);
     client.deleteObject(bucketName,filePath);
   }
